requirements.txt
alfpy==1.0.6
asgiref @ file:///opt/conda/conda-bld/asgiref_1661287532135/work
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beniget @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/beniget/beniget-0.4.1
biopython==1.81
Bottleneck @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/Bottleneck/Bottleneck-1.3.2
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164373/work
CacheControl==0.12.11
certifi==2023.5.7
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046055389/work
chardet==3.0.4
charset-normalizer==3.1.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.0.7
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1637687020823/work
cycler==0.11.0
Cython==0.29.34
deap @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/deap/deap-1.3.1
decorator==5.1.1
Django @ file:///opt/conda/conda-bld/django_1661296381686/work
djangorestframework==3.14.0
ete3==3.1.3
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1678703645500/work
executing==1.2.0
fonttools==4.39.3
gast @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/gast/gast-0.5.2
GDAL==3.3.2
h11==0.8.1
h2==3.2.0
h5py==3.8.0
hdmedians==0.14.2
hpack==3.0.0
hyperframe==5.2.0
idna==2.10
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682176699712/work
importlib-resources==5.12.0
iniconfig @ file:///home/conda/feedstock_root/build_artifacts/iniconfig_1673103042956/work
install==1.3.5
ipython==8.13.2
jedi==0.18.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1663332044897/work
kiwisolver==1.4.4
lockfile==0.12.2
lxml @ file:///croot/lxml_1679646450407/work
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpi4py @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/mpi4py/mpi4py-3.1.1
mpmath @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/mpmath/mpmath-1.2.1
msgpack==1.0.5
natsort==8.3.1
numexpr @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/numexpr/numexpr-2.7.3
numpy @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/numpy/numpy-1.21.3
packaging==23.1
pandas @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/pandas/pandas-1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1683850015520/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1667232663820/work
ply @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/ply/ply-3.11
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1679580333621/work
prompt-toolkit==3.0.38
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pybind11 @ file:///data1/apps/.local/easybuild/build/pybind11/2.7.1/GCCcore-11.2.0/pybind11-2.7.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments==2.15.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytest==7.3.1
python-dateutil==2.8.2
pythran @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/pythran/pythran-0.10.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
requests==2.30.0
rfc3986==1.5.0
scikit-bio==0.5.8
scikit-learn @ file:///croot/scikit-learn_1680198742267/work
scipy @ file:///data1/apps/.local/easybuild/build/SciPybundle/2021.10/foss-2021b/scipy/scipy-1.7.1
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse @ file:///croot/sqlparse_1668793487206/work
stack-data==0.6.2
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1644342247877/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
traitlets==5.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2023.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
